+++ backend/src/main/java/com/woowacourse/moamoa/review/domain/Reviewer.java
@@ -2,15 +2,15 @@ package com.woowacourse.moamoa.review.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
 
 @Embeddable
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Reviewer {
 
     @Column(name = "member_id", nullable = false)
     private Long memberId;
-
-    protected Reviewer() {
-    }
 
     public Reviewer(final Long memberId) {
         this.memberId = memberId;
