+++ backend/src/main/java/com/woowacourse/moamoa/study/schedule/AutoCloseEnrollmentTask.java
@@ -1,10 +1,8 @@
-package com.woowacourse.moamoa.study.batch;
+package com.woowacourse.moamoa.study.schedule;
 
 import com.woowacourse.moamoa.study.service.StudyService;
 import java.time.ZoneId;
-import org.springframework.scheduling.config.CronTask;
 import org.springframework.scheduling.config.TriggerTask;
-import org.springframework.scheduling.support.CronExpression;
 import org.springframework.scheduling.support.CronTrigger;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
