+++ backend/src/test/java/com/woowacourse/acceptance/study/AutoCloseEnrollmentAcceptanceTest.java
@@ -2,25 +2,13 @@ package com.woowacourse.acceptance.study;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.Matchers.contains;
-import static org.mockito.BDDMockito.given;
 
 import com.woowacourse.acceptance.AcceptanceTest;
-import com.woowacourse.moamoa.auth.service.oauthclient.response.GithubProfileResponse;
-import com.woowacourse.moamoa.study.service.DateTimeSystem;
-import com.woowacourse.moamoa.study.service.request.CreatingStudyRequest;
-import io.restassured.RestAssured;
-import java.time.LocalDateTime;
 import java.util.Set;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.scheduling.annotation.ScheduledAnnotationBeanPostProcessor;
 import org.springframework.scheduling.config.ScheduledTask;
 import org.springframework.scheduling.config.Task;
